requirements.txt
Jinja2==3.1.3 
MarkupSafe==2.1.5
Werkzeug==3.0.2 
blinker==1.8.0 
click==8.1.7 
colorama==0.4.6
flask==3.0.3 
itsdangerous==2.2.0 
zipp==3.18.1
boto3==1.34.93 
botocore==1.34.93 
jmespath==1.0.1 
python-dateutil==2.9.0
s3transfer==0.10.1 
six==1.16.0 
urllib3==1.26.18
dnspython==2.6.1 
pymongo==4.7.0
numpy==1.26.4
python-dotenv==1.0.1